requirements.txt
torch==2.8.0
torchvision
open-clip-torch>=2.24.0
faiss-gpu-cu12
pandas>=1.5.0
numpy>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0
scipy>=1.10.0
scikit-learn>=1.2.0
